커튼월창 UI 작업중

+++ main/contentslist/List_ConstructionCW.cs
@@ -20,7 +20,7 @@ namespace main.contentslist
         String CWNum;
         double CountDB;
         int SelectRow;
-        DataTable WindowList = new DataTable();
+        DataTable CWList = new DataTable();
         DataGridViewCheckBoxColumn checkBoxColumn = new DataGridViewCheckBoxColumn();
 
         public List_ConstructionCW()
@@ -44,7 +44,10 @@ namespace main.contentslist
         private void Add_button_Click(object sender, EventArgs e)
         {
             CWNum = Program.UTIL.CreateNum("ConstructionCW", "번호", "CW");
-            Load_form(CWNum, "Add");
+
+            Program.getMenuForm().ResetForm(6);
+
+            Load_form(CWNum, "Add", "Main");
         }
 
         public static bool OnLoadProc(Form form)
@@ -58,6 +61,7 @@ namespace main.contentslist
             }
             else if (inEditing == "Copy")
             {
+                f.LoadData(currentID);
                 f.CopyForm(currentID);
             }
             else
@@ -68,11 +72,18 @@ namespace main.contentslist
             return true;
         }
 
-        private void Load_form(String ID, String editing)
+        private void Load_form(String ID, String editing, String MainSub)
         {
             currentID = ID;
             inEditing = editing;
-            Program.getMenuForm().DoLoadForm(2, OnLoadProc);
+            if (MainSub == "Main")
+            {
+                Program.getMenuForm().DoLoadForm(2, OnLoadProc);
+            }
+            else
+            {
+             //   Program.getMenuForm().DoLoadForm(32, Sub_OnLoadProc);
+            }
         }
 
 
@@ -83,15 +94,15 @@ namespace main.contentslist
             checkBoxColumn.HeaderText = "선택";
             checkBoxColumn.Name = "check";
             dataGridView1.Columns.Add(checkBoxColumn);
-            WindowList.Columns.Add("번호", typeof(string));
-            WindowList.Columns.Add("창호 명칭", typeof(string));
-            WindowList.Columns.Add("Type", typeof(string));
-            WindowList.Columns.Add("유효열관류율" + Environment.NewLine + "[W/m²·K]", typeof(string));
-            WindowList.Columns.Add("태양열취득률" + Environment.NewLine + "[-]", typeof(string));
-            WindowList.Columns.Add("빛투과율" + Environment.NewLine + "[-]", typeof(string));
-            WindowList.Columns.Add("면적" + Environment.NewLine + "[m²]", typeof(string));
-            WindowList.Columns.Add("유리종류", typeof(string));
-            dataGridView1.DataSource = WindowList;
+            CWList.Columns.Add("번호", typeof(string));
+            CWList.Columns.Add("창호 명칭", typeof(string));
+            CWList.Columns.Add("Type", typeof(string));
+            CWList.Columns.Add("유효열관류율" + Environment.NewLine + "[W/m²·K]", typeof(string));
+            CWList.Columns.Add("태양열취득률" + Environment.NewLine + "[-]", typeof(string));
+            CWList.Columns.Add("빛투과율" + Environment.NewLine + "[-]", typeof(string));
+            CWList.Columns.Add("면적" + Environment.NewLine + "[m²]", typeof(string));
+            CWList.Columns.Add("유리종류", typeof(string));
+            dataGridView1.DataSource = CWList;
 
 
         }
@@ -100,14 +111,26 @@ namespace main.contentslist
         {
             string[][] List = Program.DB.getValue(DB.type.ProjDB, "ConstructionCW", "번호,창호명칭,Type,창호유효열관류율,태양열취득률,빛투과율,창호면적,유리종류", "");
             List<object> mainMenu = new List<object>(); // 예시 코드: 메인 메뉴 동적 할당
-
-            WindowList.Rows.Clear();
+            String Blank = "";
+            CWList.Rows.Clear();
             for (int n = 0; n < List.Length; n++)
             {
-                WindowList.Rows.Add(List[n][0], List[n][1], List[n][2], String.Format("{0:F2}", Convert.ToDouble(List[n][3])), String.Format("{0:F2}", Convert.ToDouble(List[n][4])), String.Format("{0:F2}", Convert.ToDouble(List[n][5])), String.Format("{0:F2}", Convert.ToDouble(List[n][6])), List[n][7]);
-                mainMenu.Add(new { text = List[n][1], id = "{\\\"formID\\\":6,\\\"ID\\\":\\\"" + List[n][0] + "\\\"}" }); // 예시 코드: 메인 메뉴 동적 할당
+                CWList.Rows.Add(List[n][0], Blank, List[n][1], List[n][2], Blank, String.Format("{0:F2}", Convert.ToDouble(List[n][3])), String.Format("{0:F2}", Convert.ToDouble(List[n][4])), Blank, List[n][5]);
+
+               // string[][] SubList = Program.DB.getValue(DB.type.ProjDB, "SubWindow", "번호,명칭,창호유효열관류율,창호면적", "상위창호번호 = '" + List[n][0] + "'");
+
+                List<object> subMenu = new List<object>(); // 예시 코드: 메인 메뉴 동적 할당
+
+               // for (int k = 0; k < SubList.Length; k++)
+              //  {
+
+                  //  CWList.Rows.Add(Blank, SubList[k][0], SubList[k][1], List[n][2], String.Format("{0:F2}", Convert.ToDouble(SubList[k][2])), String.Format("{0:F2}", Convert.ToDouble(List[n][3])), String.Format("{0:F2}", Convert.ToDouble(List[n][4])), String.Format("{0:F2}", Convert.ToDouble(SubList[k][3])), List[n][5]);
+                   // subMenu.Add(new { text = SubList[k][0], id = "{\\\"formID\\\":31,\\\"ID\\\":\\\"" + SubList[k][0] + "\\\"}" }); // 예시 코드: 메인 메뉴 동적 할당
+
+              //  }
+                mainMenu.Add(new { text = List[n][0] + "." + List[n][1], id = "{\\\"formID\\\":6,\\\"ID\\\":\\\"" + List[n][0] + "\\\"}", children = subMenu.ToArray() }); // 예시 코드: 메인 메뉴 동적 할당
             }
-            dataGridView1.DataSource = WindowList;
+            dataGridView1.DataSource = CWList;
             CountDB = List.Length;
             Program.UTIL.resetMainTree(1, 4, mainMenu.ToArray(), "2"); // 예시 코드: 메인 메뉴 동적 할당
         }
@@ -148,12 +171,22 @@ namespace main.contentslist
             {
                 if (k > -1)
                 {
-                    String Delete_WinNum = dataGridView1.Rows[k].Cells[1].Value.ToString();
-                    Program.DB.deleteValue(DB.type.ProjDB, "ConstructionCW", "번호 ='" + Delete_WinNum + "'");
-                    load_List();
-                }
-            }
+                    if (dataGridView1.Rows[k].Cells[1].Value.ToString() != "")
+                    {
+                        String Delete_WinNum = dataGridView1.Rows[k].Cells[1].Value.ToString();
+                        Program.DB.deleteValue(DB.type.ProjDB, "ConstructionCW", "번호 ='" + Delete_WinNum + "'");
+                        load_List();
+                    }
+                    else
+                    {
+                      //  String Delete_WinNum = dataGridView1.Rows[k].Cells[2].Value.ToString();
+                     //   Program.DB.deleteValue(DB.type.ProjDB, "SubWindow", "번호 ='" + Delete_WinNum + "'");
+                     //   load_List();
 
+                    }
+                }
+
+            }
         }
 
         private void dataGridView1_DoubleClick(object sender, EventArgs e)
@@ -161,15 +194,36 @@ namespace main.contentslist
             int k = dataGridView1.CurrentCell.RowIndex;
             if (k > -1)
             {
-                Load_form(dataGridView1.Rows[k].Cells[1].Value.ToString(), "Edit");
-            }
+                if (dataGridView1.Rows[k].Cells[1].Value.ToString() != "")
+                {
+                    Load_form(dataGridView1.Rows[k].Cells[1].Value.ToString(), "Edit", "Main");
+                }
+                else
+                {
+                  //  Load_form(dataGridView1.Rows[k].Cells[2].Value.ToString(), "Edit", "Sub");
+                }
 
+            }
         }
 
         private void Copy_button_Click(object sender, EventArgs e)
         {
-            CWNum = Program.UTIL.CreateNum("ConstructionCW", "번호", "WIN");
-            Load_form(CWNum, "Copy");
+            CWNum = Program.UTIL.CreateNum("ConstructionCW", "번호", "CW");
+            int k = dataGridView1.CurrentCell.RowIndex;
+            if (k > -1)
+            {
+                String Copy_WinNum = dataGridView1.Rows[k].Cells[1].Value.ToString();
+                if (dataGridView1.Rows[k].Cells[1].Value.ToString() != "")
+                {
+                    Program.DB.CopyValue(DB.type.ProjDB, "ConstructionCW", "번호 ='" + Copy_WinNum + "'", CWNum);
+                    Load_form(CWNum, "Copy", "Main");
+                    Load_form(CWNum, "Edit", "Main");
+                }
+                else
+                {
+                    MessageBox.Show("메인 커튼월창만 복사 가능합니다.");
+                }
+            }
         }
     }
 }
